+++ src/columns/user_login.rs
@@ -10,8 +10,8 @@ const UID_NOT_SET: u32 = 0xffffffff;
 pub struct UserLogin {
     header: String,
     unit: String,
-    fmt_contents: HashMap<i32, String>,
-    raw_contents: HashMap<i32, String>,
+    fmt_contents: HashMap<i64, String>,
+    raw_contents: HashMap<i64, String>,
     width: usize,
 }
